Move common method to utils class
